+++ client/src/components/GamePage.js
@@ -1,37 +1,27 @@
 import React, { useState } from 'react';
+import SignupForm from './SignupForm';
 
 const GamePage = () => {
   const [gameState, setGameState] = useState(1);
-  const [userName, setUsername] = useState('');
-  const [submissionError, setSubmissionError] = useState('');
+  const [currentUser, setCurrentUser] = useState('');
 
-  const onSubmit = e => {
-    e.preventDefault();
-    setSubmissionError('');
-    if (userName.length < 10 && userName.length > 1) {
+  const userInfoSubmitted = username => {
+    {
       setGameState(2);
-      return;
+      setCurrentUser(username);
     }
-    setSubmissionError('A Username between the length of 1 - 9 characters is required.');
   };
 
   return (
     <div className="game">
       {gameState === 1 && (
         <div className="game-form">
-          <form onSubmit={e => onSubmit(e)}>
-            <input
-              type="text"
-              onChange={e => setUsername(e.target.value)}
-              value={userName}
-              aria-label="Enter Username"
-              placeholder="Enter a username"
-            ></input>
-            <input type="submit" value="Submit"></input>
-            {submissionError && <label className="error-label">{submissionError}</label>}
-          </form>
+          <SignupForm
+            onFormSubmitted={username => userInfoSubmitted(username)}
+          ></SignupForm>
         </div>
       )}
+      {gameState === 2 && <h1>{currentUser}</h1>}
     </div>
   );
 };
